services/quantum-dwave/requirements.txt
# Quantum D-Wave Service - Requirements
# Python 3.10+

# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# D-Wave Ocean SDK
dwave-ocean-sdk==6.7.0
dimod==0.12.14
neal==0.6.0
dwave-system==1.23.0
dwave-preprocessing==0.6.5

# Scientific Computing
numpy==1.24.3
pandas==2.0.3
networkx==3.0
scipy==1.11.4

# Utilities
python-multipart==0.0.6
httpx==0.25.1

# Monitoring
prometheus-client==0.18.0
python-json-logger==2.0.7

# Environment
python-dotenv==1.0.0
